Add clear command

+++ src/lib/creds-helper.ts
@@ -4,6 +4,7 @@ import { exec } from 'child_process';
 import { STS } from 'aws-sdk';
 import { ICredentials, IUserIdentity, IProfile } from './interfaces';
 import { isProfile } from './profile-helper';
+import Command from '@oclif/command';
 const ini = require('ini');
 
 export async function initCredentials(profile: string = 'default'): Promise<IUserIdentity> {
@@ -78,6 +79,19 @@ export function writeCredentialsFile(credentials: ICredentials) {
   parsedCredentials.default.aws_access_key_id = credentials.accessKeyId;
   parsedCredentials.default.aws_secret_access_key = credentials.secretAccessKey;
   parsedCredentials.default.aws_session_token = credentials.sessionToken;
+  const encodedCredentials = ini.encode(parsedCredentials);
+  writeFileSync(credentialsFilePath, encodedCredentials, {encoding: 'utf-8'});
+}
+
+export function clearCredentials(command: Command) {
+  const credentialsFilePath = `${homedir()}/.aws/credentials`;
+  if (!existsSync(credentialsFilePath)) {
+    command.log('> credentials file does not exit');
+  }
+  const parsedCredentials = ini.parse(readFileSync(credentialsFilePath, 'utf-8'));
+  parsedCredentials.default.aws_access_key_id = '';
+  parsedCredentials.default.aws_secret_access_key = '';
+  parsedCredentials.default.aws_session_token = '';
   const encodedCredentials = ini.encode(parsedCredentials);
   writeFileSync(credentialsFilePath, encodedCredentials, {encoding: 'utf-8'});
 }
